backend/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.3.1
Deprecated==1.3.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Limiter==4.1.1
Flask-RESTful==0.3.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.6.0
MarkupSafe==3.0.3
ordered-set==4.1.0
packaging==25.0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
typing_extensions==4.15.0
Werkzeug==3.1.4
wrapt==2.0.1
